+++ CHANGELOG.md
@@ -11,6 +11,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 * Fixed lineHeight behavior for `Konva.Text`
 * Some performance optimizations for `Konva.Text`
 * Better wrap algorithm for `Konva.Text`
+* fixed `Konva.Arrow` with tension != 0
 
 ## [2.0.3][2018-04-21]